requirements.txt
open3d
opencv-python-headless==4.7.0.72
PyYAML==6.0
tqdm==4.64.0
matplotlib
scipy
ipython
jupyter


# --extra-index-url https://download.pytorch.org/whl/cu113
# torch==1.12.1+cu113
# torchvision==0.13.1+cu113
# torchaudio==0.12.1
# torchgeometry==0.1.2
torch==1.12.1
transforms3d
torchgeometry
tensorboardX
pyntcloud
mediapy
pytest